Modify isVisible to take shadows into account
Models can cast shadows into the scene even when the models
themselves are entirely outside of the viewport.
This commit deobfuscates the isVisible function, and extends it
when shadow rendering is enabled such that models are considered
visible also when their shadow is visible.

Note, it works by assuming the shadow lands on the lowest part of
the scene bounds, and will therefore sometimes include models that
do not cast shadows into the scene. Improving this requires further
scene analysis.